Conflicts:
src/library/features/playlist/playlisttablemodel.cpp
src/library/features/playlist/playlisttablemodel.h
src/library/basesqltablemodel.h

+++ src/library/features/recording/dlgrecording.h
@@ -30,7 +30,7 @@ class DlgRecording : public QFrame, public Ui::DlgRecording {
   public slots:
     void toggleRecording(bool toggle);
     void slotRecordingEnabled(bool);
-    void slotBytesRecorded(long);
+    void slotBytesRecorded(int);
     void refreshBrowseModel();
     void slotDurationRecorded(QString durationRecorded);
